Add files via upload
I mistakenly did everything for this app in the tlelab folder the walkthrough had me create instead of in the download conversation-simple-master folder. Because of this, I went into the necessary manifest file and .env file and filled out the correct information. This is now the same as the tlelab folder I used to host my app on bluemix.